# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/tenant-cv/me": {
      "get": {
        "operationId": "TenantCvController_getMyCv",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tenant CV"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/tenant-cv/current": {
      "get": {
        "operationId": "TenantCvController_getCurrentCv",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tenant CV"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/tenant-cv": {
      "put": {
        "operationId": "TenantCvController_updateMyCv",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTenantCvDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tenant CV"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/tenant-cv/complete": {
      "post": {
        "operationId": "TenantCvController_completeMyCv",
        "summary": "Mark tenant onboarding as completed (idempotent)",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Completion recorded"
          }
        },
        "tags": [
          "Tenant CV"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/tenant-cv/share": {
      "post": {
        "operationId": "TenantCvController_createShareLink",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Tenant CV"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/tenant-cv/{userId}/verification": {
      "patch": {
        "operationId": "TenantCvController_setVerification",
        "summary": "Set KYC/referencing badges on a tenant CV (admin)",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetVerificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification badges updated"
          }
        },
        "tags": [
          "Tenant CV"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/tenant-cv/{share_uuid}": {
      "get": {
        "operationId": "TenantCvController_getPublicCv",
        "parameters": [
          {
            "name": "share_uuid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tenant CV"
        ]
      }
    },
...
  },
  "info": {
    "title": "TA-DA! Rental Platform API",
    "description": "API for connecting tenants and property operators",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "Bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer. Пример: 'Bearer 12345abcdef'",
        "name": "Authorization",
        "type": "http",
        "in": "Header"
      }
    },
    "schemas": {
      "RentHistoryEntryDto": {
        "type": "object",
        "properties": {
          "property_name": {
            "type": "string",
            "description": "Property name",
            "example": "Kings Cross Apartments"
          },
          "address": {
            "type": "string",
            "description": "Property address line",
            "example": "37 Swinton Street"
          },
          "city": {
            "type": "string",
            "description": "City or area",
            "example": "London"
          },
          "price_per_month": {
            "type": "number",
            "description": "Monthly rent price",
            "example": 1712
          },
          "bedrooms": {
            "type": "number",
            "description": "Bedrooms count",
            "example": 1
          },
          "bathrooms": {
            "type": "number",
            "description": "Bathrooms count",
            "example": 1
          },
          "size_sqft": {
            "type": "number",
            "description": "Size in sqft",
            "example": 497
          },
          "property_type": {
            "type": "string",
            "description": "Property type",
            "example": "apartment"
          },
          "furnishing": {
            "type": "string",
            "description": "Furnishing description",
            "example": "furnished"
          },
          "match_score": {
            "type": "number",
            "description": "Matching score percent",
            "example": 90
          },
          "review": {
            "type": "string",
            "description": "Landlord or concierge review",
            "example": "Excellent tenant"
          },
          "landlord": {
            "type": "string",
            "description": "Landlord/author name",
            "example": "King Cross Apartments"
          },
          "period_from": {
            "type": "string",
            "description": "Start date ISO",
            "example": "2021-06-01"
          },
          "period_to": {
            "type": "object",
            "description": "End date ISO or null",
            "example": "2024-03-01"
          },
          "media_url": {
            "type": "string",
            "description": "Optional media preview URL",
            "example": "https://..."
          }
        }
      },
      "UpdateTenantCvDto": {
        "type": "object",
        "properties": {
          "headline": {
            "type": "string",
            "description": "Custom headline for CV (e.g., Ready to move - Long term 6+ m)"
          },
          "about_me": {
            "type": "string",
            "description": "About me text to show on the CV"
          },
          "hobbies": {
            "description": "Custom hobbies list (overrides preferences.hobbies on CV)",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "rent_history": {
            "description": "Rent history entries to display on CV",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RentHistoryEntryDto"
            }
          },
          "share_uuid": {
            "type": "string",
            "description": "Optional existing share UUID to preserve (ignored on update)"
          }
        }
      },
      "SetVerificationDto": {
        "type": "object",
        "properties": {
          "kyc_status": {
            "type": "string",
            "description": "KYC (identity) verification state",
            "enum": [
              "not_started",
              "in_progress",
              "passed",
              "failed"
            ],
            "example": "passed"
          },
          "referencing_status": {
            "type": "string",
            "description": "Referencing verification state",
            "enum": [
              "not_started",
              "in_progress",
              "passed",
              "failed"
            ],
            "example": "in_progress"
          }
        }
      },
...
    }
  }
}